Moving Changeset extensions to correct namespace
